Rename netconf.sal.connect.api to netconf.client.mdsal.api
https://github.com/opendaylight/netconf/commit/8b6271ebe2a4e745e66b940b668fe905892deb0e

JIRA: LIGHTY-235
